requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
djangorestframework==3.14.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
